requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
Flask==2.0.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
pkgutil-resolve-name==1.3.10
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.5
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zipp==3.9.0
